# api/openapi.yaml
openapi: 3.1.0
info:
  title: kbdb API
  description: >
    API for tracking a personal mechanical keyboard collection: keyboards,
    switches, keycap sets, and the builds that combine them. Routes are
    nested under /v1/users/{userId}/... to support both a caller's own
    collection and items other users have explicitly shared with them —
    see the userId path parameter's description for the exact rules.
  version: "1.0.0"
  contact:
    url: https://mykeebs.info
  license:
    name: MIT
    url: https://github.com/rogueserenity/kbdb/blob/main/LICENSE
servers:
  - url: https://api.mykeebs.info
security:
  - BearerAuth: []
...
paths:
  /v1/users/{userId}/keyboards:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [keyboards]
      summary: List keyboards
      operationId: listKeyboards
      security:
        - {}
        - BearerAuth: []
      description: >
        If userId is the caller's own subject, lists all of the caller's keyboards. Otherwise, lists only that user's keyboards explicitly shared with the caller (possibly empty).
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of keyboards.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyboardListPage"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    post:
      tags: [keyboards]
      summary: Create a keyboard
      operationId: createKeyboard
      description: >
        userId must be the caller's own subject; creating in another user's collection always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeyboardInput"
      responses:
        "201":
          description: The created keyboard.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyboard"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
  /v1/users/{userId}/keycap-sets:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [keycap-sets]
      summary: List keycap sets
      operationId: listKeycapSets
      security:
        - {}
        - BearerAuth: []
      description: >
        If userId is the caller's own subject, lists all of the caller's keycap sets. Otherwise, lists only that user's keycap sets explicitly shared with the caller (possibly empty).
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: A page of keycap sets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeycapSetListPage"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    post:
      tags: [keycap-sets]
      summary: Create a keycap set
      operationId: createKeycapSet
      description: >
        userId must be the caller's own subject; creating in another user's collection always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeycapSetInput"
      responses:
        "201":
          description: The created keycap set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeycapSet"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /v1/users/{userId}/keycap-sets/{keycapSetId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/KeycapSetId"
    get:
      tags: [keycap-sets]
      summary: Get a keycap set
      operationId: getKeycapSet
      security:
        - {}
        - BearerAuth: []
      description: >
        Returns the keycap set if it belongs to the caller, or if its owner has shared it with the caller. Returns 404 (not 403) if the keycap set does not exist, is not owned by the caller, and is not shared with the caller — this avoids revealing that an unshared item exists.
      responses:
        "200":
          description: The requested keycap set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeycapSet"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    put:
      tags: [keycap-sets]
      summary: Update a keycap set
      operationId: updateKeycapSet
      description: >
        userId must be the caller's own subject; updating another user's keycap set always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeycapSetInput"
      responses:
        "200":
          description: The updated keycap set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeycapSet"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    delete:
      tags: [keycap-sets]
      summary: Delete a keycap set
      operationId: deleteKeycapSet
      description: >
        userId must be the caller's own subject; deleting another user's
        keycap set always returns 404. on_delete controls what happens if
        any kit in the set is still referenced by a build: "block" (the
        default) fails with 409 and lists the blocking build ids;
        "cascade" deletes the set and every build that references any kit
        in it, returning their ids; "detach" deletes the set regardless,
        leaving referencing builds with a dangling keycap_kits[] entry.
      parameters:
        - name: on_delete
          in: query
          schema:
            type: string
            enum: [block, cascade, detach]
            default: block
          description: How to handle a set still referenced by one or more builds.
      responses:
        "200":
          description: Deleted via cascade; lists the builds that were also deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CascadeDeleteResult"
        "204":
          description: Deleted (block with no references, or detach).
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          description: >
            Blocked by one or more referencing builds (on_delete=block, the
            default — the Problem body has blocking_build_ids).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /v1/users/{userId}/keycap-sets/{keycapSetId}/kits:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/KeycapSetId"
    post:
      tags: [keycap-sets]
      summary: Add a kit to a keycap set
      operationId: createKeycapKit
      description: >
        userId must be the caller's own subject; adding a kit to another
        user's keycap set always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeycapKitInput"
      responses:
        "201":
          description: The created kit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeycapKit"
        "400":
          $ref: "#/components/responses/ProblemResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
  /v1/users/{userId}/keycap-sets/{keycapSetId}/kits/{kitId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
      - $ref: "#/components/parameters/KeycapSetId"
      - name: kitId
        in: path
        required: true
        schema:
          type: string
        description: The kit_id returned when the kit was created.
    put:
      tags: [keycap-sets]
      summary: Update a kit
      operationId: updateKeycapKit
      description: >
        userId must be the caller's own subject; updating a kit on another
        user's keycap set always returns 404.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeycapKitInput"
      responses:
        "200":
          description: The updated kit.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeycapKit"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
        default:
          $ref: "#/components/responses/ProblemResponse"
    delete:
      tags: [keycap-sets]
      summary: Delete a kit
      operationId: deleteKeycapKit
      description: >
        userId must be the caller's own subject; deleting a kit from
        another user's keycap set, or from a keycap set that doesn't exist,
        both return 404. Idempotent — deleting a kitId not present in the
        set is not an error. on_delete controls what happens if the kit is
        still referenced by a build: "block" (the default) fails with 409
        and lists the blocking build ids; "cascade" deletes the kit and
        every build that references it, returning their ids; "detach"
        deletes the kit regardless, leaving referencing builds with a
        dangling keycap_kits[] entry. A 409 can also mean the set is being
        modified concurrently (unrelated to on_delete) — distinguish the
        two by whether the Problem body has a blocking_build_ids field.
      parameters:
        - name: on_delete
          in: query
          schema:
            type: string
            enum: [block, cascade, detach]
            default: block
          description: How to handle a kit still referenced by one or more builds.
      responses:
        "200":
          description: Deleted via cascade; lists the builds that were also deleted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CascadeDeleteResult"
        "204":
          description: Deleted (block with no references, or detach).
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "409":
          description: >
            Either blocked by one or more referencing builds (on_delete=block, the default —
            the Problem body has blocking_build_ids), or the set is being modified
            concurrently (no blocking_build_ids; retry).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        default:
          $ref: "#/components/responses/ProblemResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An IdP-issued OIDC JWT (Authorization Bearer <token>).
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
      description: >
        The IdP subject of the item's owner — the `sub` claim of an access
        token, and the `user_id` field of a profile. If this matches the
        authenticated caller's own subject, the request is against their own
        collection. If it names a different user, the request is against
        that user's collection: reads return only items that user has
        explicitly shared with the caller (or, for a specific item, only if
        that item is shared with the caller — otherwise a 404, not a 403,
        to avoid revealing that an item exists but isn't shared); writes
        (create/update/delete) always fail with a 404 in this case, since a
        caller can never modify another user's items. Not a PII/security
        concern to expose in a URL — an IdP subject is an opaque identifier,
        not derived from email or any other identifying attribute.
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return.
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response's next_cursor.
  responses:
    ProblemResponse:
      description: An error, formatted per RFC 9457 (Problem Details for HTTP APIs).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
...
    KeyboardListPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/KeyboardSummary"
        next_cursor:
          type: [string, "null"]
...